+++ distributed/shuffle/tests/utils.py
@@ -1,0 +1,45 @@
+from __future__ import annotations
+
+import itertools
+from typing import Any
+
+from distributed.core import PooledRPCCall
+from distributed.shuffle._shuffle import ShuffleId
+from distributed.shuffle._worker_extension import ShuffleRun
+
+
+class PooledRPCShuffle(PooledRPCCall):
+    def __init__(self, shuffle: ShuffleRun):
+        self.shuffle = shuffle
+
+    def __getattr__(self, key):
+        async def _(**kwargs):
+            from distributed.protocol.serialize import nested_deserialize
+
+            method_name = key.replace("shuffle_", "")
+            kwargs.pop("shuffle_id", None)
+            kwargs.pop("run_id", None)
+            # TODO: This is a bit awkward. At some point the arguments are
+            # already getting wrapped with a `Serialize`. We only want to unwrap
+            # here.
+            kwargs = nested_deserialize(kwargs)
+            meth = getattr(self.shuffle, method_name)
+            return await meth(**kwargs)
+
+        return _
+
+
+class AbstractShuffleTestPool:
+    _shuffle_run_id_iterator = itertools.count()
+
+    def __init__(self, *args, **kwargs):
+        self.shuffles = {}
+
+    def __call__(self, addr: str, *args: Any, **kwargs: Any) -> PooledRPCShuffle:
+        return PooledRPCShuffle(self.shuffles[addr])
+
+    async def shuffle_barrier(self, id: ShuffleId, run_id: int) -> dict[str, None]:
+        out = {}
+        for addr, s in self.shuffles.items():
+            out[addr] = await s.inputs_done()
+        return out
